+++ src/test/java/gameObjects/goTest.java
@@ -8,6 +8,6 @@ public class goTest {
     public static void main(String[] args) {
         Renderer.init();
         GameLoop.start();
-        WorldUpdater.addObject(new TestGO(Renderer.unitsWide / 2, Renderer.unitsTall / 2, 2, 2));
+        WorldUpdater.addObject(new TestGO((int) (Renderer.unitsWide / 2), (int) (Renderer.unitsTall / 2), 2, 2));
     }
 }
